pyproject.toml
[tool.poetry]
name = "httpxhar"
version = "0.1.0"
description = "Convert httpx Request to curl command"
authors = ["Congee <congeec@qq.com>"]

[tool.poetry.dependencies]
python = "^3.7"
httpx = "^0.11.1"

[tool.poetry.dev-dependencies]
flake8-isort = "^2.8.0"
pytest-asyncio = "^0.10.0"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
